Restructuring files
moving the header files to an include folder
